DataOffload: Move and separate tests for new sub-package
