+++ loki/transformations/data_offload/tests/sources/projGlobalVarImports/moduleA.F90
