+++ loki/transformations/data_offload/tests/sources/projGlobalVarImports/moduleB.F90
