+++ loki/transformations/data_offload/tests/sources/projGlobalVarImports/moduleC.F90
